Verifying body
